Switch price_per_hour and agreed_price to float
`int * duration // 3600` truncates early when the per-hour rate is small
and the duration is sub-hour - `1 * 1800 // 3600 = 0` even though the
correct amount is 0.5. Switching to float carries that intermediate
precision and only rounds once at the wei boundary:
`int(price_per_hour * duration_seconds / 3600)`.

Touches the full negotiation chain: Pydantic schemas (price_per_hour,
agreed_price, initial_price, max_price, proposed_price, etc.), the
codec's amount formula, the BisectionStrategy midpoint
(`(our + their) // 2` -> `/ 2`), int(...) casts on price values become
float(...), and the persistence-port signatures match.

Token decimals are now a presentation concern (CLI Decimal rendering)
- the wire layer no longer bakes "integer x 10^decimals" into the
schema.

+++ buyer/market_buyer/aggregation.py
@@ -319,7 +319,7 @@ async def gather_outcomes(
     return await asyncio.gather(*(_one(c) for c in candidates))
 
 
-def _extract_advertised_price(match: dict[str, Any]) -> int | None:
+def _extract_advertised_price(match: dict[str, Any]) -> float | None:
     """Pull the per-hour advertised price from a match's first accepted escrow.
 
     Mirrors what the seller advertises: ``accepted_escrows[0].price_per_hour``
@@ -339,7 +339,7 @@ def _extract_advertised_price(match: dict[str, Any]) -> int | None:
         return None
     amount = first.get("price_per_hour")
     try:
-        parsed = int(amount) if amount is not None else None
+        parsed = float(amount) if amount is not None else None
     except (ValueError, TypeError):
         return None
     if parsed is None or parsed <= 0:
