+++ storefront/src/market_storefront/models/negotiation_models.py
@@ -21,7 +21,7 @@ class NegotiateNewRequest(BaseModel):
 
     listing_id: str
     buyer_address: str
-    initial_price: int = Field(ge=0)
+    initial_price: float = Field(ge=0)
     provision_terms: ProvisionTerms
     escrow_proposal: EscrowProposal
     buyer_agent_url: str = ""
@@ -39,7 +39,7 @@ class NegotiateNewResponse(BaseModel):
 
     negotiation_id: str
     action: str
-    price: int | None = None
+    price: float | None = None
     reason: str | None = None
     accepted_provision_terms: ProvisionTerms | None = None
     accepted_escrow_proposal: EscrowProposal | None = None
@@ -48,13 +48,13 @@ class NegotiateNewResponse(BaseModel):
 class NegotiateContinueRequest(BaseModel):
     action: Literal["counter", "accept", "exit"]
     buyer_address: str
-    price: int | None = None
+    price: float | None = None
     reason: str | None = None
 
 
 class NegotiateContinueResponse(BaseModel):
     action: str
-    price: int | None = None
+    price: float | None = None
     reason: str | None = None
 
 
@@ -63,7 +63,7 @@ class NegotiationSummary(BaseModel):
     our_listing_id: str
     their_agent_id: str | None = None
     terminal_state: str | None = None
-    agreed_price: int | None = None
+    agreed_price: float | None = None
     round_count: int = 0
     created_at: str | None = None
     model_config = {"extra": "allow"}
@@ -81,7 +81,7 @@ class NegotiationMessage(BaseModel):
     round: int
     sender: str
     action_taken: str
-    proposed_price: int | None = None
+    proposed_price: float | None = None
     model_config = {"extra": "allow"}
 
 
@@ -90,7 +90,7 @@ class NegotiationDetailResponse(BaseModel):
     our_listing_id: str
     their_agent_id: str | None = None
     terminal_state: str | None = None
-    agreed_price: int | None = None
+    agreed_price: float | None = None
     round_count: int = 0
     messages: list[dict[str, Any]] = Field(default_factory=list)
     stage_events: list[dict[str, Any]] = Field(default_factory=list)
@@ -100,21 +100,21 @@ class NegotiationDetailResponse(BaseModel):
 
 class AdvanceRequest(BaseModel):
     action: Literal["counter", "accept", "exit"]
-    price: int | None = None
+    price: float | None = None
     reason: str | None = None
 
 
 class ForceAcceptRequest(BaseModel):
-    price: int
+    price: float
 
 
 class ForceAcceptResponse(BaseModel):
     action: str
-    price: int
+    price: float
     source: str = "admin_force_accept"
 
 
 class AdvanceResponse(BaseModel):
     action: str
-    price: int | None = None
+    price: float | None = None
     reason: str | None = None
